Move mergeXml method to util module
