+++ cordova-lib/spec-cordova/prepare.spec.js
@@ -87,7 +87,7 @@ describe('prepare command', function() {
         load = spyOn(lazy_load, 'based_on_config').andReturn(Q());
         cp = spyOn(shell, 'cp').andReturn(true);
         mkdir = spyOn(shell, 'mkdir');
-        spyOn(prepare, '_mergeXml');
+        spyOn(util, 'mergeXml');
         spyOn(ConfigParser.prototype, 'write');
         spyOn(xmlHelpers, 'parseElementtreeSync').andCallFake(function() {
             return new et.ElementTree(et.XML(TEST_XML));
@@ -176,99 +176,5 @@ describe('prepare command', function() {
                 }).fin(done);
             });
         });
-    });
-});
-
-describe('prepare._mergeXml', function () {
-    var dstXml;
-    beforeEach(function() {
-        dstXml = et.XML(TEST_XML);
-    });
-    it('should merge attributes and text of the root element without clobbering', function () {
-        var testXml = et.XML('<widget foo="bar" id="NOTANID">TEXT</widget>');
-        prepare._mergeXml(testXml, dstXml);
-        expect(dstXml.attrib.foo).toEqual('bar');
-        expect(dstXml.attrib.id).not.toEqual('NOTANID');
-        expect(dstXml.text).not.toEqual('TEXT');
-    });
-
-    it('should merge attributes and text of the root element with clobbering', function () {
-        var testXml = et.XML('<widget foo="bar" id="NOTANID">TEXT</widget>');
-        prepare._mergeXml(testXml, dstXml, 'foo', true);
-        expect(dstXml.attrib.foo).toEqual('bar');
-        expect(dstXml.attrib.id).toEqual('NOTANID');
-        expect(dstXml.text).toEqual('TEXT');
-    });
-
-    it('should not merge platform tags with the wrong platform', function () {
-        var testXml = et.XML('<widget><platform name="bar"><testElement testAttrib="value">testTEXT</testElement></platform></widget>'),
-            origCfg = et.tostring(dstXml);
-
-        prepare._mergeXml(testXml, dstXml, 'foo', true);
-        expect(et.tostring(dstXml)).toEqual(origCfg);
-    });
-
-    it('should merge platform tags with the correct platform', function () {
-        var testXml = et.XML('<widget><platform name="bar"><testElement testAttrib="value">testTEXT</testElement></platform></widget>'),
-            origCfg = et.tostring(dstXml);
-
-        prepare._mergeXml(testXml, dstXml, 'bar', true);
-        expect(et.tostring(dstXml)).not.toEqual(origCfg);
-        var testElement = dstXml.find('testElement');
-        expect(testElement).toBeDefined();
-        expect(testElement.attrib.testAttrib).toEqual('value');
-        expect(testElement.text).toEqual('testTEXT');
-    });
-
-    it('should merge singelton children without clobber', function () {
-        var testXml = et.XML('<widget><author testAttrib="value" href="http://www.nowhere.com">SUPER_AUTHOR</author></widget>');
-
-        prepare._mergeXml(testXml, dstXml);
-        var testElements = dstXml.findall('author');
-        expect(testElements).toBeDefined();
-        expect(testElements.length).toEqual(1);
-        expect(testElements[0].attrib.testAttrib).toEqual('value');
-        expect(testElements[0].attrib.href).toEqual('http://cordova.io');
-        expect(testElements[0].attrib.email).toEqual('dev@cordova.apache.org');
-        expect(testElements[0].text).toContain('Apache Cordova Team');
-    });
-
-    it('should clobber singelton children with clobber', function () {
-        var testXml = et.XML('<widget><author testAttrib="value" href="http://www.nowhere.com">SUPER_AUTHOR</author></widget>');
-
-        prepare._mergeXml(testXml, dstXml, '', true);
-        var testElements = dstXml.findall('author');
-        expect(testElements).toBeDefined();
-        expect(testElements.length).toEqual(1);
-        expect(testElements[0].attrib.testAttrib).toEqual('value');
-        expect(testElements[0].attrib.href).toEqual('http://www.nowhere.com');
-        expect(testElements[0].attrib.email).toEqual('dev@cordova.apache.org');
-        expect(testElements[0].text).toEqual('SUPER_AUTHOR');
-    });
-
-    it('should append non singelton children', function () {
-        var testXml = et.XML('<widget><preference num="1"/> <preference num="2"/></widget>');
-
-        prepare._mergeXml(testXml, dstXml, '', true);
-        var testElements = dstXml.findall('preference');
-        expect(testElements.length).toEqual(4);
-    });
-
-    it('should handle namespaced elements', function () {
-        var testXml = et.XML('<widget><foo:bar testAttrib="value">testText</foo:bar></widget>');
-
-        prepare._mergeXml(testXml, dstXml, 'foo', true);
-        var testElement = dstXml.find('foo:bar');
-        expect(testElement).toBeDefined();
-        expect(testElement.attrib.testAttrib).toEqual('value');
-        expect(testElement.text).toEqual('testText');
-    });
-
-    it('should not append duplicate non singelton children', function () {
-        var testXml = et.XML('<widget><preference name="fullscreen" value="true"/></widget>');
-
-        prepare._mergeXml(testXml, dstXml, '', true);
-        var testElements = dstXml.findall('preference');
-        expect(testElements.length).toEqual(2);
     });
 });
